requirements.txt
pypdf
spacy
sentence-transformers
scikit-learn
typer[all]
pypdf
docx2txt
python-dotenv
sentence-transformers>=2.6.0
